base Score logic - need refactoring;

+++ src/com/triple/game/elements/ElementType.java
@@ -113,7 +113,7 @@ public class ElementType {
 		this.subspecies = subspecies;
 	}
 
-	public double getScore() {
+	public int getScore() {
 		return score;
 	}
